Conflicts:
	svd.py

+++ svd.py
@@ -63,8 +63,8 @@ def init_features():
     global movie_features
 
     for f in xrange(NUM_FEATURES):
-        user_features[f] = np.array([0.1 for i in xrange(NUM_USERS)])
-        movie_features[f] = np.array([0.1 for i in xrange(NUM_MOVIES)])
+        user_features[f] = np.array([0.1 for i in xrange(NUM_USERS)], dtype=np.float16)
+        movie_features[f] = np.array([0.1 for i in xrange(NUM_MOVIES)], dtype=np.float16)
 
 # Takes either mu or um numpy array (from data.npz or data-mu.npz), returns a
 # list of lists, index being movie/user index, value being movie avg or user
